+++ MetaExchange.ConsoleApp/Models/ExecutionOrder.cs
@@ -5,7 +5,7 @@ namespace MetaExchange.ConsoleApp.Models
         public string ExchangeId { get; set; }
         public string OrderId { get; set; }
         public string Type { get; set; } // "Buy" or "Sell"
-        public double Amount { get; set; }
-        public double Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
     }
 }
