+++ MetaExchange.ConsoleApp/Models/ExecutionPlan.cs
@@ -5,6 +5,6 @@ namespace MetaExchange.ConsoleApp.Models
     public class ExecutionPlan
     {
         public List<ExecutionOrder> Orders { get; set; } = new List<ExecutionOrder>();
-        public double TotalCostOrRevenue { get; set; }
+        public decimal TotalCostOrRevenue { get; set; }
     }
 }
